chore: sett BehandlingskvalitetV2 til gjeldende versjon
