+++ apps/skde/src/components/IndicatorTable/chartrow/DataQualityPopup.tsx
